% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignDataDS.R
\name{assignDataDS}
\alias{assignDataDS}
\title{assign data to one or multiple servers with some encrypted data from the analysis computer}
\usage{
assignDataDS(
  header = "",
  payload = "",
  property.a = 0,
  property.b = 0,
  property.c = 0,
  property.d = 0
)
}
\description{
This server function assigns some values into a specific structure.
}
