% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decryptParamDS.R
\name{decryptParamDS}
\alias{decryptParamDS}
\title{decrypt a server parameter}
\usage{
decryptParamDS(param_names = NULL)
}
\arguments{
\item{param_names}{character argument. Name of the variable used  to store the parameter value on a server.}
}
\description{
This server function decrypts a given parameter in matrix.
}
