% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decryptParamDS.R
\name{deParamDS}
\alias{deParamDS}
\alias{decryptParamDS}
\title{decrypt a server parameter}
\usage{
decryptParamDS(param.name = NULL)
}
\description{
This server function decrypts a given parameter in matrix.
}
