% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeParamDS.R
\name{encodeParameterDS}
\alias{encodeParameterDS}
\alias{encodeParamDS}
\title{encode a server parameter}
\usage{
encodeParamDS(param.name = NULL)
}
\description{
This server function encodes a given parameter using a dot product and two shared secrets.
}
