% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataDS.R
\name{getDataDS}
\alias{getDataDS}
\title{Retrieves the encrypted data from a server to the analysis computer}
\usage{
getDataDS(master_mode = TRUE)
}
\description{
This server function retrieves some encrypted data to be passed onto the analysis computer
}
