% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEncodedDataDS.R
\name{getEncodedDataDS}
\alias{getEncodedDataDS}
\title{Retrieves the encoded data}
\usage{
getEncodedDataDS()
}
\description{
This server function retrieves some encoded data to be passed onto the analysis computer
}
