% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateExchangeDS.R
\name{initiateExchangeDS}
\alias{initiateExchangeDS}
\title{initiates the first steps of a parameter in a master server}
\usage{
initiateExchangeDS(master = TRUE)
}
\description{
This server function creates the data required to exchange a parameter securely between two
DataSHIELD server.
}
