% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEncodedDataDS.R
\name{sendEncodedDataDS}
\alias{sendEncodedDataDS}
\title{decodes some data and transform them into a matrix}
\usage{
sendEncodedDataDS(
  header = "",
  payload = "",
  property.a = 0,
  property.b = 0,
  property.c = 0,
  property.d = 0
)
}
\description{
This server function assigns some values into a specific structure.
}
