% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setReceivedMatrix.R
\name{setReceivedMatrixDS}
\alias{setReceivedMatrixDS}
\title{decodes some data and transform them into a matrix}
\usage{
setReceivedMatrixDS(data = "", no.column = 0)
}
\description{
This server function assigns some values into a specific structure.
}
